requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
google-cloud-firestore